refactored data-types into one file
