--- src/data-types/page.ts
@@ -1,7 +1,0 @@
-import { GraphQLInputFieldConfigMap, GraphQLInt } from 'graphql'
-import { memoizeContextModel } from '../utils'
-
-export const page = memoizeContextModel(() => ({
-  limit: { type: GraphQLInt },
-  offset: { type: GraphQLInt },
-}))
