--- src/data-types/where.ts
@@ -1,16 +1,0 @@
-import { GraphQLInputFieldConfigMap, isType } from 'graphql'
-import { memoizeContextModel, reduceContextFields } from '../utils'
-
-export const where = memoizeContextModel(buildModeModel =>
-  reduceContextFields(
-    buildModeModel,
-    buildModeModel.baseFilters(),
-    (where, attr, type, field) => ({
-      ...where,
-      ...buildModeModel.buildMode.filterStrategy(
-        isType(type) ? type : field,
-        attr.name,
-      ),
-    }),
-  ),
-)
